[Ref] Silence warning.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@16148 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Snd_fx.cpp
@@ -5959,7 +5959,7 @@ uint32 CSoundFile::GetPeriodFromNote(uint32 note, int32 nFineTune, uint32 nC5Spe
 		{
 			// Compute everything in Hertz rather than periods.
 			uint32 freq = Util::muldiv_unsigned(nC5Speed, LinearSlideUpTable[(note % 12u) * 16u] << (note / 12u), 65536 << 5);
-			LimitMax(freq, int32_max);
+			LimitMax(freq, static_cast<uint32>(int32_max));
 			return freq;
 		} else if(m_SongFlags[SONG_LINEARSLIDES])
 		{
